+++ src/components/index.tsx
@@ -27,6 +27,7 @@ import RouterEnter from './auth/RouterEnter';
 import Cssmodule from './cssmodule';
 import MapUi from './ui/map';
 import QueryParams from './extension/QueryParams';
+import Visitor from './extension/Visitor';
 
 const WysiwygBundle = Loadable({
     // 按需加载富文本配置
@@ -59,4 +60,5 @@ export default {
     Cssmodule,
     MapUi,
     QueryParams,
+    Visitor,
 } as any;