+++ src/utils/index.ts
@@ -19,3 +19,10 @@ export const queryString = () => {
     });
     return _queryString;
 };
+
+/**
+ * 校验是否登录
+ * @param permits
+ */
+export const checkLogin = (permits: any): boolean =>
+    (process.env.NODE_ENV === 'production' && !!permits) || process.env.NODE_ENV === 'development';